Emma

+++ resources/views/layouts/app.blade.php
@@ -33,6 +33,7 @@
     <!-- INTERNAL Switcher css -->
     <link href="{{ asset('admin') }}/switcher/css/switcher.css" rel="stylesheet">
     <link href="{{ asset('admin') }}/switcher/demo.css" rel="stylesheet">
+    <link href="https://cdnjs.cloudflare.com/ajax/libs/jquery-toast-plugin/1.3.2/jquery.toast.min.css" rel="stylesheet">
 
 </head>
 
@@ -108,6 +109,8 @@
 
         <!-- Switcher js -->
         <script src="{{ asset('admin') }}/switcher/js/switcher.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery-toast-plugin/1.3.2/jquery.toast.min.js"></script>
         <script>
             class DataTable {
                 constructor(tableId, url, columns) {
@@ -125,7 +128,7 @@
                     });
                 }
 
-                create(data,url) {
+                create(data,url,errorFunction) {
                     // Mengirim data ke backend untuk melakukan create
                     // Setelah berhasil, melakukan refresh pada tabel
                     $.ajax({
@@ -135,6 +138,23 @@
                         success: () => {
                             $(`#${this.tableId}`).DataTable().ajax.reload();
                             $('#modal-form').modal('hide');
+                            // menghapus class 'is-invalid' pada inputan
+                            $('.is-invalid').removeClass('is-invalid');
+                            // menghapus elemen <span> dengan class 'error'
+                            $('.error').remove();
+                            toast('Data berhasil di tambahkan','success')
+                        },
+                        error: (xhr, status, error) => {
+                            toast('Telah terjadi kesalahan','error')
+                            $('.is-invalid').removeClass('is-invalid');
+                            // menghapus elemen <span> dengan class 'error'
+                            $('.error').remove();
+                            $.each(xhr.responseJSON.errors, function (key, value) {
+                                let escapedKey = key.replace(/\./g, '\\.');
+                                let inputan = $(`input#${escapedKey}`);
+                                inputan.addClass("is-invalid");
+                                inputan.parent().append("<span class='error text-danger'>" + value[0] + "</span>");
+                            });
                         }
                     });
                 }
@@ -149,7 +169,6 @@
                             // Menampilkan data pada modal atau form
                             $('#modal').modal('show');
                             $('#name').val(data.name);
-                            // ...
                         }
                     });
                 }
@@ -164,22 +183,66 @@
                         success: () => {
                             $(`#${this.tableId}`).DataTable().ajax.reload();
                             $('#modal-form').modal('hide');
+                            // menghapus class 'is-invalid' pada inputan
+                            $('.is-invalid').removeClass('is-invalid');
+                            // menghapus elemen <span> dengan class 'error'
+                            $('.error').remove();
+                            toast('Data berhasil di perbarui','success')
                         }
                     });
                 }
 
-                delete(id) {
+                delete(url) {
                     // Mengirim data ke backend untuk melakukan delete
                     // Setelah berhasil, melakukan refresh pada tabel
-                    $.ajax({
-                        url: '',
-                        method: 'DELETE',
-                        data: { id: id },
-                        success: () => {
-                            $(`#${this.tableId}`).DataTable().ajax.reload();
+                    Swal.fire({
+                        title: 'Apa kamu yakin?',
+                        text: "Anda tidak akan dapat mengembalikan ini!",
+                        icon: 'warning',
+                        showCancelButton: true,
+                        confirmButtonColor: '#3085d6',
+                        cancelButtonColor: '#d33',
+                        confirmButtonText: 'Ya, hapus!',
+                        cancelButtonText: 'Batal',
+                        }).then((result) => {
+                        if (result.isConfirmed) {
+                            $.ajax({
+                                url: url,
+                                method: 'DELETE',
+                                success: () => {
+                                    $(`#${this.tableId}`).DataTable().ajax.reload();
+                                    Swal.fire(
+                                        'Dihapus!',
+                                        'Data Anda telah dihapus.',
+                                        'success'
+                                    )
+                                }
+                            });
+                            
                         }
-                    });
+                    })
+                    
                 }
+            }
+            function toast(text,icon){
+                $.toast({
+                    text: text, // Text that is to be shown in the toast
+                    heading: 'Note', // Optional heading to be shown on the toast
+                    icon: icon, // Type of toast icon
+                    showHideTransition: 'fade', // fade, slide or plain
+                    allowToastClose: true, // Boolean value true or false
+                    hideAfter: 3000, // false to make it sticky or number representing the miliseconds as time after which toast needs to be hidden
+                    stack: 5, // false if there should be only one toast at a time or a number representing the maximum number of toasts to be shown at a time
+                    position: 'bottom-left', // bottom-left or bottom-right or bottom-center or top-left or top-right or top-center or mid-center or an object representing the left, right, top, bottom values
+                    
+                    textAlign: 'left',  // Text alignment i.e. left, right or center
+                    loader: true,  // Whether to show loader or not. True by default
+                    loaderBg: '#9EC600',  // Background color of the toast loader
+                    beforeShow: function () {}, // will be triggered before the toast is shown
+                    afterShown: function () {}, // will be triggered after the toat has been shown
+                    beforeHide: function () {}, // will be triggered before the toast gets hidden
+                    afterHidden: function () {}  // will be triggered after the toast has been hidden
+                });
             }
         </script>
         @yield('js')
